+++ Assets/Debug/DebugScene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3409a33eb178df9499e8b2863f14ada1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
