+++ Assets/Debug/DebugScene/IMGPREFAB.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ab998dca4e3cb6440bb3fe2a32c8ed82
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
