+++ Assets/Debug/DebugScene/Latest.cs
@@ -1,0 +1,17 @@
+﻿using UnityEngine;
+
+namespace MDMulti_DEBUG.DebugScene
+{
+    public class Latest : MonoBehaviour
+    {
+        public void Latest1()
+        {
+            Debug.Log("L");
+        }
+
+        public void Latest2()
+        {
+            Debug.Log("L2");
+        }
+    }
+}
