+++ Assets/Debug/DebugScene/Latest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60ca37e232af0d047980570be24db55e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
