+++ Assets/Debug/DebugScene/PeerConnection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d131638568a99646a8bea9f1c8d6ed9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
