+++ Assets/Debug/DebugScene/Rest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7213adcf17e7c564c9f5c5abb75e460b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
