+++ Assets/Debug/DebugScene/SecureClient.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1643191e020d984bad85109226f1aa3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
