+++ Assets/Debug/DebugScene/TEXTPREFAB.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2e8e46b2076394b499aaa4d49bc28218
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
