+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,6 +9,6 @@ EditorBuildSettings:
     path: 
     guid: 00000000000000000000000000000000
   - enabled: 1
-    path: Assets/Debug/DebugScene.unity
+    path: Assets/Debug/DebugScene/DebugScene.unity
     guid: d10fe52a8642e9a49bc34899993b1a92
   m_configObjects: {}
